Move the Start Lock from JVMCloudService to DefaultCloudService
